Rename module `Crucible.JVM.Class` to `Crucible.JVM.Translation.Class`.
