+++ crucible-jvm/src/Lang/Crucible/JVM.hs
@@ -10,17 +10,17 @@ Stability        : provisional
 module Lang.Crucible.JVM
   ( module Lang.Crucible.JVM.Types
   , module Lang.Crucible.JVM.Context
-  , module Lang.Crucible.JVM.Class
   , module Lang.Crucible.JVM.Overrides
   , module Lang.Crucible.JVM.Translation
+  , module Lang.Crucible.JVM.Translation.Class
   , module Lang.Crucible.JVM.Simulate
   , module Lang.Crucible.JVM.ClassRefs
   ) where
 
 import Lang.Crucible.JVM.Types
 import Lang.Crucible.JVM.Context
-import Lang.Crucible.JVM.Class
 import Lang.Crucible.JVM.Overrides
 import Lang.Crucible.JVM.Translation
+import Lang.Crucible.JVM.Translation.Class
 import Lang.Crucible.JVM.Simulate
 import Lang.Crucible.JVM.ClassRefs
