+++ crucible-jvm/src/Lang/Crucible/JVM/Overrides.hs
@@ -99,7 +99,7 @@ import           Lang.Crucible.JVM.Types
 import           Lang.Crucible.JVM.Context
 import           Lang.Crucible.JVM.ClassRefs
 import           Lang.Crucible.JVM.Translation.Monad
-import           Lang.Crucible.JVM.Class
+import           Lang.Crucible.JVM.Translation.Class
 
 import Debug.Trace
