+++ crucible-jvm/src/Lang/Crucible/JVM/Translation/Class.hs
@@ -1,6 +1,6 @@
 {- |
 
-Module           : Lang.Crucible.JVM.Class
+Module           : Lang.Crucible.JVM.Translation.Class
 Description      : Implements OO features of the JVM
 Copyright        : (c) Galois, Inc 2018
 License          : BSD3
@@ -46,7 +46,7 @@ working with objects, dynamic class information and arrays.
 
 {-# OPTIONS_GHC -haddock #-}
 
-module Lang.Crucible.JVM.Class
+module Lang.Crucible.JVM.Translation.Class
    (
      lookupClassGen
    , getAllFields
